fix atualização coment, processo

+++ src/templates/CardEA/index.js
@@ -1,27 +1,32 @@
-import React from 'react';                
+import React, { useState } from 'react';                
 import api from '../../Services/api';
 
 export default function CardEA ({process}) {
-;
+    const [verifica, setVerifica] = useState('');
+
     async function updateProcess() {
-        if(window.confirm('Deseja mesmo aceitar o processo?')) {
-            await api.put(`process/atualiza/${process.idProcesso}`);
-            // eslint-disable-next-line no-restricted-globals
-            location.reload();
-        }
-        else {
-            alert('Erro ao atualizar processo. Tente novamente.');
-        }
+        setVerifica(true);
     }
     async function deleteProcess() {
-        if(window.confirm('Deseja mesmo deletar o processo?')) {
-            await api.delete(`process/deleta/${process.idProcesso}`);
+        setVerifica(false);
+    }
 
-            // eslint-disable-next-line no-restricted-globals
-            window.location.reload();
+    async function atualiza() {
+        if(verifica === true) {
+            if(window.confirm('Deseja mesmo aceitar o processo?')) {
+                await api.put(`process/atualiza/${process.idProcesso}`);
+            }
+            else {
+                alert('Erro ao atualizar processo. Tente novamente.');
+            }
         }
-        else {
-            alert('Erro ao deletar processo. Tente novamente.');
+        else if (verifica === false) {
+            if(window.confirm('Deseja mesmo deletar o processo?')) {
+                await api.delete(`process/deleta/${process.idProcesso}`);
+            }
+            else {
+                alert('Erro ao deletar processo. Tente novamente.');
+            }
         }
     }
 
@@ -31,10 +36,12 @@ export default function CardEA ({process}) {
                 <p>Lar temporário: {process.userNameLT}</p>
                 <p>Status: {process.status}</p>
 
-                <div>
-                    <button className="buttonsProcess" onClick={updateProcess}>aceitar</button>
-                    <button className="buttonsProcess" onClick={deleteProcess}>negar</button>
-                </div>
+                <form onSubmit={atualiza}>
+                    <div>
+                        <button type='submit' className="buttonsProcess" onClick={updateProcess}>aceitar</button>
+                        <button type='submit' className="buttonsProcess" onClick={deleteProcess}>negar</button>
+                    </div>
+                </form>
             </li>
     );
 }
